More refactoring of ExtensionManagerViewModel
