+++ src/extensibility/ExtensionManagerView.js
@@ -28,168 +28,13 @@
 define(function (require, exports, module) {
     "use strict";
     
-    var Strings                = require("strings"),
-        NativeApp              = require("utils/NativeApp"),
-        ExtensionManager       = require("extensibility/ExtensionManager"),
-        InstallExtensionDialog = require("extensibility/InstallExtensionDialog"),
-        registry_utils         = require("extensibility/registry_utils"),
-        itemTemplate           = require("text!htmlContent/extension-manager-view-item.html");
-    
-    /**
-     * @private
-     * @type {Array}
-     * A list of fields to search when trying to search for a query string in an object. Each field is 
-     * represented as an array of keys to recurse downward through the object. We store this here to avoid 
-     * doing it for each search call.
-     */
-    var _searchFields = [["metadata", "name"], ["metadata", "title"], ["metadata", "description"],
-                         ["metadata", "author", "name"], ["metadata", "keywords"], ["owner"]];
-    /**
-     * @constructor
-     * The model for the ExtensionManagerView. Keeps track of the extensions that are currently visible
-     * and manages sorting/filtering them.
-     * Events:
-     *     filterChange - triggered whenever the filtered set changes (including on initialize).
-     */
-    function Model() {
-    }
-    
-    /**
-     * @type {Object}
-     * The current set of extensions managed by this model. The keys are the IDs of the 
-     * extensions, or, for legacy extensions with no ID, the full local path to the extension.
-     * Depending on whether a given extension is in the registry or not and whether it's locally 
-     * installed or not, each entry might or might not have the following contents:
-     *     metadata: the contents of the extension's package.json. Available for both local and registry extensions.
-     *     owner: the ID of the owner who uploaded the extension. Only available for extensions that are in the registry.
-     *     versions: the history of versions of the extension. Only available for extensions that are in the registry.
-     *     status: the current status of the extension. Only available for local extensions.
-     *     path: the full file path of the extension. Only available for local extensions.
-     *     installType: one of "default", "dev", "user", or "unknown". Only available for local extensions.
-     */
-    Model.prototype.extensions = null;
-    
-    /**
-     * @type {Array.<Object>}
-     * The list of IDs of items matching the current query and sorted with the current sort.
-     */
-    Model.prototype.filterSet = null;
-    
-    /**
-     * @type {Object}
-     * The list of all ids from the extension list, sorted with the current sort.
-     */
-    Model.prototype._sortedFullSet = null;
-    
-    /**
-     * @private
-     * @type {string}
-     * The last query we filtered by. Used to optimize future searches.
-     */
-    Model.prototype._lastQuery = null;
-    
-    /**
-     * @private
-     * Sets up the initial filtered set based on the sorted full set.
-     */
-    Model.prototype._setInitialFilter = function () {
-        // Initial filtered list is the same as the sorted list.
-        this.filterSet = this._sortedFullSet.slice(0);
-        $(this).triggerHandler("filterChange");
-    };
-    
-    /**
-     * Initializes the model from the remote extension registry.
-     * @return {$.Promise} a promise that's resolved with the registry JSON data
-     * or rejected if the server can't be reached.
-     */
-    Model.prototype.initializeFromRegistry = function () {
-        var self = this;
-        return ExtensionManager.getRegistry(true)
-            .done(function (registry) {
-                self.extensions = registry;
-                
-                // Sort the registry by last published date and store the sorted list of IDs.
-                self._sortedFullSet = registry_utils.sortRegistry(registry).map(function (entry) {
-                    return entry.metadata.name;
-                });
-                self._setInitialFilter();
-            });
-    };
-    
-    /**
-     * Initializes the model from the set of locally installed extensions, sorted
-     * alphabetically by id (or name of the extension folder for legacy extensions).
-     */
-    Model.prototype.initializeFromInstalledExtensions = function () {
-        this.extensions = ExtensionManager.loadedExtensions;
-        this._sortedFullSet = Object.keys(this.extensions)
-            .map(function (key) {
-                var match = key.match(/\/([^\/]+)$/);
-                return (match && match[1]) || key;
-            })
-            .sort();
-        this._setInitialFilter();
-    };
-    
-    /**
-     * @private
-     * Searches for the given query in the current extension list and updates the filter set,
-     * dispatching a filterChange event.
-     * @param {string} query The string to search for.
-     */
-    Model.prototype.filter = function (query) {
-        var self = this, initialList, newFilterSet = [];
-        if (this._lastQuery && query.indexOf(this._lastQuery) === 0) {
-            // This is the old query with some new letters added, so we know we can just
-            // search in the current filter set.
-            initialList = this.filterSet;
-        } else {
-            // This is a new query, so start with the full list.
-            initialList = this._sortedFullSet;
-        }
-        
-        query = query.toLowerCase();
-        initialList.forEach(function (id) {
-            var entry = self.extensions[id];
-            if (entry && self._entryMatchesQuery(entry, query)) {
-                newFilterSet.push(id);
-            }
-        });
-        
-        this._lastQuery = query;
-        this.filterSet = newFilterSet;
-        $(this).trigger("filterChange");
-    };
-    
-    /**
-     * @private
-     * Tests if the given entry matches the query.
-     * @param {Object} entry The extension entry to test.
-     * @param {string} query The query to match against.
-     * @return {boolean} Whether the query matches.
-     */
-    Model.prototype._entryMatchesQuery = function (entry, query) {
-        return _searchFields.some(function (fieldSpec) {
-            var i, cur = entry;
-            for (i = 0; i < fieldSpec.length; i++) {
-                // Recurse downward through the specified fields to the leaf value.
-                cur = cur[fieldSpec[i]];
-                if (!cur) {
-                    return false;
-                }
-            }
-            // If the leaf value is an array (like keywords), search each item, otherwise
-            // just search in the string.
-            if (Array.isArray(cur)) {
-                return cur.some(function (keyword) {
-                    return keyword.toLowerCase().indexOf(query) !== -1;
-                });
-            } else if (cur.toLowerCase().indexOf(query) !== -1) {
-                return true;
-            }
-        });
-    };
+    var Strings                   = require("strings"),
+        NativeApp                 = require("utils/NativeApp"),
+        ExtensionManagerViewModel = require("extensibility/ExtensionManagerViewModel").ExtensionManagerViewModel,
+        ExtensionManager          = require("extensibility/ExtensionManager"),
+        registry_utils            = require("extensibility/registry_utils"),
+        InstallExtensionDialog    = require("extensibility/InstallExtensionDialog"),
+        itemTemplate              = require("text!htmlContent/extension-manager-view-item.html");
     
     /**
      * @constructor
@@ -199,7 +44,7 @@ define(function (require, exports, module) {
      */
     function ExtensionManagerView() {
         var self = this;
-        this.model = new Model();
+        this.model = new ExtensionManagerViewModel();
         this._itemTemplate = Mustache.compile(itemTemplate);
         this.$el = $("<div class='extension-list'/>");
         this._$table = $("<table class='table'/>").appendTo(this.$el);
@@ -207,7 +52,7 @@ define(function (require, exports, module) {
         // Show the busy spinner and access the registry.
         var $spinner = $("<div class='spinner large spin'/>")
             .appendTo(this.$el);
-        this.model.initializeFromRegistry().done(function () {
+        this.model.initialize(ExtensionManagerViewModel.EXTENSIONS_REGISTRY).done(function () {
             self._setupEventHandlers();
             self._render();
         }).fail(function () {
@@ -266,6 +111,7 @@ define(function (require, exports, module) {
         // Listen for extension status changes.
         $(ExtensionManager).on("statusChange", function (e, id, status) {
             // Re-render the extension item.
+            // TODO: should this flow through the model?
             // FUTURE: later on, some of these views might be for installed extensions that aren't 
             // in the registry, e.g. legacy extensions or local dev extensions.
             var extensions = self.model.extensions;
@@ -357,7 +203,4 @@ define(function (require, exports, module) {
     };
     
     exports.ExtensionManagerView = ExtensionManagerView;
-    
-    // For unit test only
-    exports.Model = Model;
 });
